Moved value parsers to their own file
